Move pull request to separate package
